requirements.txt
tornado~=6.1
nbformat~=5.3.0
ipyreact~=0.3.0
ipython~=8.5.0
traitlets~=5.3.0
python-dotenv~=1.0.0
watchdog~=3.0.0
pandas>=2.0.0
numpy>=1.0.0